refactor : Gnb 컴포넌트 Logo 분리
