+++ src/app/(user)/challenges/[challengeId]/create/page.jsx
@@ -1,10 +1,13 @@
 import Container from "@/components/container/PageContainer";
+import Logo from "@/layout/_components/Logo";
 import React from "react";
 
 export default function page() {
   return (
-    <Container>
-      <div>신규 작업물 생성 페이지</div>
+    <Container maxWidth="max-w-4xl" className="bg-brand-yellow">
+      <header className="flex items-center justify-between">
+        <Logo />
+      </header>
     </Container>
   );
 }
