+++ src/app/page.jsx
@@ -1,11 +1,16 @@
-import Container, {RootLandingPageContainer} from "@/components/container/PageContainer";
-import Gnb from "@/layout/_components/Gnb";
-
+import Container, {
+  RootLandingPageContainer,
+} from "@/components/container/PageContainer";
+import Gnb from "@/layout/Gnb";
 
 export default function Home() {
   // 랜딩에선 따로 헤더 넣어주기
   return (
-    <Container maxWidth="w-full" padding="p-0 sm:p-0 lg:p-0" className="flex flex-col">
+    <Container
+      maxWidth="w-full"
+      padding="p-0 sm:p-0 lg:p-0"
+      className="flex flex-col"
+    >
       <Gnb userRole="guest" />
       <div>랜딩페이지</div>
       {/* <Container maxWidth="w-full" className="bg-brand-black flex flex-col items-center" >
