+++ src/layout/_components/Logo.jsx
@@ -1,0 +1,18 @@
+import React from "react";
+import logo from "@/assets/img/img_logo.svg";
+import Link from "next/link";
+import Image from "next/image";
+
+export default function Logo() {
+  return (
+    <Link href="/">
+      <Image
+        src={logo}
+        alt="Docthur 로고"
+        width={80}
+        height={18}
+        className="h-[18px] w-[80px] sm:h-[27px] sm:w-[120px]"
+      />
+    </Link>
+  );
+}
